+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Pending
+
+* [Load V5] Extract base class for Campaign Manager [#13684](https://github.com/Applifier/unity-ads-webview/pull/13684)
+* [Load V5] Remove Unused Methods in LegacyCampaignManager [#13714](https://github.com/Applifier/unity-ads-webview/pull/13714)
+
+# Wed Mar 11 20:05:18 UTC 2020
+
+* Add getCurrentExperiment to MediationLoadTrackingManager [#13687](https://github.com/Applifier/unity-ads-webview/pull/13687)
+* Open Measurement - Fix omid doubleclick - Admob Integration [#13713](https://github.com/Applifier/unity-ads-webview/pull/13713)
+
 # Mon Mar  9 20:49:40 UTC 2020
 
 * Run Open Measurement For Double Click [#13511](https://github.com/Applifier/unity-ads-webview/pull/13511)